Refactor: Enhance async support and improve method signatures across various components
- Updated BufferPool methods to support async operations, including getPage, flushPage, and evictPage.
- Modified HeapTable methods to utilize async for insert, read, delete, and update operations, ensuring better concurrency handling.
- Adjusted DatabaseServer's executeQuery method to return rows directly, improving usability.
- Enhanced SchemaEvolutionManager to ensure proper type handling for estimatedCompletion and duration.
- Refactored QueryOptimizerManager to return [[PlanNode]] instead of [[QueryNode]] for consistency.
- Updated transaction management methods to handle async operations, improving overall performance and responsiveness.

+++ Sources/ColibriCore/Catalog/SchemaEvolution.swift
@@ -510,7 +510,7 @@ public actor SchemaEvolutionManager {
             phase: .preparation,
             progress: 0,
             startTime: currentTime,
-            estimatedCompletion: currentTime + schemaChangeTimeout,
+            estimatedCompletion: currentTime + UInt64(schemaChangeTimeout),
             blockingOperations: []
         )
         
@@ -559,7 +559,7 @@ public actor SchemaEvolutionManager {
             ddlStatement: "ONLINE \(currentChange.changeType.rawValue)",
             executedBy: txId,
             executedAt: currentTime,
-            duration: currentTime - currentChange.startTime,
+            duration: Int(currentTime - currentChange.startTime),
             success: true,
             rollbackScript: ""
         )
